// apps/exam-vault/src/generated/models/section-model.ts
import type { Exam } from './exam-model';

export const SectionSectionTypeKeyToLabel = {
  'SectionTypeKey0': 'MCQ',
  'SectionTypeKey1': 'ShortText',
  'SectionTypeKey2': 'Code',
  'SectionTypeKey3': 'Listening',
  'SectionTypeKey4': 'Reading',
  'SectionTypeKey5': 'Writing',
  'SectionTypeKey6': 'Speaking',
  'SectionTypeKey7': 'Interview'
} as const;
export type SectionSectionTypeKey = keyof typeof SectionSectionTypeKeyToLabel;

export interface Section {
  /**
   * @displayName Section ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  name1: string;
  /**
   * @displayName Exam
   * @validationRule Required for create/update operations
   */
  exam: Pick<Exam, 'id' | 'title'>;
  /**
   * @displayName Instructions
   */
  instructions?: string;
  /**
   * @displayName Media URL
   */
  mediaURL?: string;
  /**
   * @displayName Order Index
   * @validationRule Required for create/update operations
   */
  orderIndex: number;
  /**
   * @displayName Passage Text
   */
  passageText?: string;
  /**
   * @displayName Section Type
   * @validationRule Required for create/update operations
   */
  sectionTypeKey: SectionSectionTypeKey;
  /**
   * @displayName Time Limit (Minutes)
   */
  timeLimitMinutes?: number;
}

export const _Section = 'Section' as const;